User now is getting an option to run this script, in case it is accidentally ran, since it makes a lot of changes

+++ FileOrganizer.py
@@ -12,24 +12,28 @@ if not os.path.exists (currentDirectory):
     print ("Path specified does not exist.")
 
 else:
-    for subdir, dirs, files in os.walk (currentDirectory):
-        for file in files:
-            pathToFile = os.path.join (currentDirectory, file)
+    print "Are you certain you want to organize your files? (Y/N)"
+    choice = raw_input()
+    
+    if choice.upper() == 'Y':
+        for subdir, dirs, files in os.walk (currentDirectory):
+            for file in files:
+                pathToFile = os.path.join (currentDirectory, file)
 
-            if os.path.isfile (pathToFile):
-                # If file is THIS file, skip
-                if os.path.basename (__file__) != file:
-                    # Ignore hidden files
-                    if not file.startswith ('.'):
-                        # Get only text from file extension and capitalize
-                        extension = os.path.splitext (file)[1].replace ('.', '')
-                        # Check to make sure file has an exension (if not, its an empty string)
-                        if (len (extension) <= 0):
-                            extension = "MISC"
+                if os.path.isfile (pathToFile):
+                    # If file is THIS file, skip
+                    if os.path.basename (__file__) != file:
+                        # Ignore hidden files
+                        if not file.startswith ('.'):
+                            # Get only text from file extension and capitalize
+                            extension = os.path.splitext (file)[1].replace ('.', '')
+                            # Check to make sure file has an exension (if not, its an empty string)
+                            if (len (extension) <= 0):
+                                extension = "MISC"
 
-                        pathToNewDirectory = os.path.join (currentDirectory, extension)
+                            pathToNewDirectory = os.path.join (currentDirectory, extension)
 
-                        if not os.path.exists (pathToNewDirectory):
-                            os.mkdir (pathToNewDirectory)
+                            if not os.path.exists (pathToNewDirectory):
+                                os.mkdir (pathToNewDirectory)
 
-                        os.rename (pathToFile, os.path.join (pathToNewDirectory, file))
+                            os.rename (pathToFile, os.path.join (pathToNewDirectory, file))
